# Conflicts:
#	app/build.gradle
#	app/src/main/java/za/co/woolworths/financial/services/android/util/AppConstant.kt
#	app/src/main/res/values/strings.xml

+++ app/src/main/java/za/co/woolworths/financial/services/android/models/dto/cart/OrderSummary.kt
@@ -5,13 +5,13 @@ import za.co.woolworths.financial.services.android.models.dto.Suburb
 import za.co.woolworths.financial.services.android.models.dto.voucher_and_promo_code.DiscountDetails
 
 class OrderSummary {
-    var totalItemsCount: Int = 0
+    var totalItemsCount: Double = 0.0
     var groupSubTotal: String = ""
-    var giftCardAuthorizedAmount: Int = 0
+    var giftCardAuthorizedAmount: Double = 0.0
     var orderId: String = ""
     var basketTotal: Double = 0.00
     var shippingAdjusted: Boolean? = null
-    var wrewardsDiscount: Int = 0
+    var wrewardsDiscount: Double = 0.0
     var store: StoreDetails? = null
     var submittedDate: String = ""
     var completedDate: String = ""
@@ -20,9 +20,9 @@ class OrderSummary {
     var discountDetails: DiscountDetails? = null
     var suburb: Suburb? = null
     var state: String = ""
-    var estimatedDelivery: Int = 0
+    var estimatedDelivery: Double = 0.0
     var deliveryStatus: Any? = null
-    var totalOrderCount: Int = 0
-    var savedAmount: Int = 0
+    var totalOrderCount: Double = 0.0
+    var savedAmount: Double = 0.0
     var deliveryDates: Any? = null
 }
